Fix issues with context popping
Previously, adding  a new category when an existing entry was being
edited resulted in app crashes due to context not being found. The
problem seems to lie in `showDialog()` method, which is primarily used
for dialogs, not pages. This commit uses routes for navigation. As a
side-effect, animations are smoother, too.