Move sign in/registration navigation to middleware
Moves navigation related to sign in/registration to middleware in order
to simplify widgets. Note, however, that this commit does not move all
pages to the initial route configuration. Namely, pages that require
arguments will need further work before it is possible to navigate to
them using pushNamed() and similar methods.