Wait until old widget is disposed of
It seems that once a new widget is pushed, the old one is not disposed of
until the new one finishes animating in. However, while the old route is
still active, it seems that it (sometimes) receives events from
Firestore. This causes permission denied if we log out before Flutter is
done with disposing of a widget. Therefore, a hack is to wait for a bit
until we are reasonably sure that an old widget is disposed of.

There might be another way. However, so far I have not been able to find
it.